+++ tests/constants.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+BIRDS_ALARM_FILE = Path("tests") / "payloads" / "birds.wav"
+FAKE_FILE = Path("tests") / "payloads" / "file.mp4"
+NOT_SUPPORTED_ERROR = "is not supported.*.mp3, .mp4, .wav are."
